split ContentFrom function to indivisual file and create debug tag version
